Make VoxelMesh great again by reducing complexity and time needed

+++ include/voxel_mesh.h
@@ -16,7 +16,7 @@ public:
   // void drawAsCubes(Canvas &canvas);
   std::vector<float> intensity_array;
   std::vector<Vec3> position_array;
-  unsigned int column{1}, line{1}, depth{1};
+  unsigned short column{1}, line{1}, depth{1};
 
 protected:
   float offset[3]{1.0f, 1.0f, 1.0f};
